+++ crates/ruma-server-util/src/lib.rs
@@ -1,6 +1,0 @@
-#![doc(html_favicon_url = "https://ruma.dev/favicon.ico")]
-#![doc(html_logo_url = "https://ruma.dev/images/logo.png")]
-//! Collection of helpers for implementing Matrix homeservers using Ruma.
-
-#![warn(missing_docs)]
-pub mod authorization;