+++ crates/ruma-federation-api/CHANGELOG.md
@@ -5,6 +5,13 @@ Breaking changes:
 - Remove the `origin` field in `create_join_event::{v1/v2}::RoomState` due to a
   clarification in the spec.
 
+Improvements:
+
+- The `unstable-unspecified` cargo feature was removed. The `pdus` field of
+  `send_transaction_message::v1::Request` is always serialized. To allow this
+  field to be missing during deserialization, use the `compat-optional-txn-pdus`
+  cargo feature.
+
 # 0.11.0
 
 Improvements: